+++ app/Http/Controllers/PagesController.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace App\Http\Controllers;
+
+use Illuminate\Http\Request;
+
+class PagesController extends Controller
+{
+    public function index() {
+        return view('non-auth.homepage');
+    }
+}
